requirements.txt
# requirements.txt
django
djangorestframework
mysqlclient
requests
